+++ lib/msal-angular/src/msal.module.ts
@@ -23,7 +23,7 @@ export class MsalModule {
     static forRoot(
         config: Configuration,
         angularConfig: MsalAngularConfiguration = defaultMsalAngularConfiguration
-    ): ModuleWithProviders {
+    ): ModuleWithProviders<MsalModule> {
         return {
             ngModule: MsalModule,
             providers: [