+++ lib/msal-browser/src/app/IPublicClientApplication.ts
@@ -8,6 +8,7 @@ import { RedirectRequest } from "../request/RedirectRequest";
 import { PopupRequest } from "../request/PopupRequest";
 import { SilentRequest } from "../request/SilentRequest";
 import { SsoSilentRequest } from "../request/SsoSilentRequest";
+import { BrowserConfigurationAuthError } from "../error/BrowserConfigurationAuthError";
 
 export interface IPublicClientApplication {
     acquireTokenPopup(request: PopupRequest): Promise<AuthenticationResult>;
@@ -16,6 +17,7 @@ export interface IPublicClientApplication {
     addEventCallback(callback: Function): string | null;
     removeEventCallback(callbackId: string): void;
     getAccountByHomeId(homeAccountId: string): AccountInfo | null;
+    getAccountByLocalId(localId: string): AccountInfo | null;
     getAccountByUsername(userName: string): AccountInfo | null;
     getAllAccounts(): AccountInfo[];
     handleRedirectPromise(hash?: string): Promise<AuthenticationResult | null>;
@@ -26,3 +28,54 @@ export interface IPublicClientApplication {
     getLogger(): Logger;
     setLogger(logger: Logger): void;
 }
+
+export const stubbedPublicClientApplication: IPublicClientApplication = {
+    acquireTokenPopup: () => {
+        return Promise.reject(BrowserConfigurationAuthError.createStubPcaInstanceCalledError);
+    },
+    acquireTokenRedirect: () => {	
+        return Promise.reject(BrowserConfigurationAuthError.createStubPcaInstanceCalledError);	
+    },	
+    acquireTokenSilent: () => {	
+        return Promise.reject(BrowserConfigurationAuthError.createStubPcaInstanceCalledError);	
+    },	
+    getAllAccounts: () => {
+        return [];	
+    },	
+    getAccountByHomeId: () => {
+        return null;
+    },
+    getAccountByUsername: () => {	
+        return null;	
+    },	
+    getAccountByLocalId: () => {
+        return null;
+    },
+    handleRedirectPromise: () => {	
+        return Promise.reject(BrowserConfigurationAuthError.createStubPcaInstanceCalledError);	
+    },	
+    loginPopup: () => {	
+        return Promise.reject(BrowserConfigurationAuthError.createStubPcaInstanceCalledError);	
+    },	
+    loginRedirect: () => {	
+        return Promise.reject(BrowserConfigurationAuthError.createStubPcaInstanceCalledError);	
+    },	
+    logout: () => {	
+        return Promise.reject(BrowserConfigurationAuthError.createStubPcaInstanceCalledError);	
+    },	
+    ssoSilent: () => {	
+        return Promise.reject(BrowserConfigurationAuthError.createStubPcaInstanceCalledError);	
+    },
+    addEventCallback: () => {
+        return null;
+    },
+    removeEventCallback: () => {
+        return;
+    },
+    getLogger: () => {
+        throw BrowserConfigurationAuthError.createStubPcaInstanceCalledError();
+    },
+    setLogger: () => {
+        return;
+    }
+};