+++ lib/msal-common/src/request/RequestParameterBuilder.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { AADServerParamKeys, Constants, ResponseMode, SSOTypes, ClientInfo, AuthenticationScheme, ClaimsRequestKeys } from "../utils/Constants";
+import { AADServerParamKeys, Constants, ResponseMode, SSOTypes, ClientInfo, AuthenticationScheme, ClaimsRequestKeys, PasswordGrantConstants} from "../utils/Constants";
 import { ScopeSet } from "./ScopeSet";
 import { ClientConfigurationError } from "../error/ClientConfigurationError";
 import { StringDict } from "../utils/MsalTypes";
@@ -68,6 +68,23 @@ export class RequestParameterBuilder {
     }
 
     /**
+     * add post logout redirectUri
+     * @param redirectUri
+     */
+    addPostLogoutRedirectUri(redirectUri: string): void {
+        RequestValidator.validateRedirectUri(redirectUri);
+        this.parameters.set(AADServerParamKeys.POST_LOGOUT_URI, encodeURIComponent(redirectUri));
+    }
+
+    /**
+     * add id_token_hint to logout request
+     * @param idTokenHint 
+     */
+    addIdTokenHint(idTokenHint: string): void {
+        this.parameters.set(AADServerParamKeys.ID_TOKEN_HINT, encodeURIComponent(idTokenHint));
+    }
+
+    /**
      * add domain_hint
      * @param domainHint
      */
@@ -85,7 +102,7 @@ export class RequestParameterBuilder {
 
     /**
      * add sid
-     * @param sid 
+     * @param sid
      */
     addSid(sid: string): void {
         this.parameters.set(SSOTypes.SID, encodeURIComponent(sid));
@@ -222,7 +239,7 @@ export class RequestParameterBuilder {
     addClientAssertionType(clientAssertionType: string): void {
         this.parameters.set(AADServerParamKeys.CLIENT_ASSERTION_TYPE, encodeURIComponent(clientAssertionType));
     }
-    
+
     /**
      * add OBO assertion for confidential client flows
      * @param clientAssertion
@@ -230,7 +247,7 @@ export class RequestParameterBuilder {
     addOboAssertion(oboAssertion: string): void {
         this.parameters.set(AADServerParamKeys.OBO_ASSERTION, encodeURIComponent(oboAssertion));
     }
-    
+
     /**
      * add grant type
      * @param grantType
@@ -296,8 +313,24 @@ export class RequestParameterBuilder {
     }
 
     /**
+     * adds `username` for Password Grant flow
+     * @param username
+     */
+    addUsername(username: string): void {
+        this.parameters.set(PasswordGrantConstants.username, username);
+    }
+
+    /**
+     * adds `password` for Password Grant flow
+     * @param password
+     */
+    addPassword(password: string): void {
+        this.parameters.set(PasswordGrantConstants.password, password);
+    }
+
+    /**
      * add pop_jwk to query params
-     * @param cnfString 
+     * @param cnfString
      */
     addPopToken(cnfString: string): void {
         if (!StringUtils.isEmpty(cnfString)) {